added stashed changes(renaming of files and folders)
